Patch schemas in diffSchemas for subgraphs using extend keyword; merge schema def and schema extensions directives in schema diff renderer
